Made style changes in ToolsToShareOnOtherWebsites and CodeCopier components. Moved CodeCopier component.
