+++ src/sass/components/_more-tools.scss
@@ -1,0 +1,26 @@
+$margin-below-textarea: .1em 0;
+$padding-within-textarea: .5rem .7em;
+
+.textarea-clipboard {
+  -webkit-appearance: none;
+  background-color: $white;
+  border: 1px solid $gray-darker;
+  border-radius: $radius-none;
+  box-sizing: border-box;
+  color: $text-color;
+  cursor: auto;
+  display: block;
+  flex-direction: column;
+  font-size: 1rem;
+  height: 5rem;
+  line-height: 1.3;
+  margin: $margin-below-textarea;
+  max-width: 46rem;
+  outline: none;
+  padding: $padding-within-textarea;
+  resize: auto;
+  user-select: text;
+  white-space: pre-wrap;
+  width: 100%;
+  word-wrap: break-word;
+}
