+++ src/sass/main.scss
@@ -92,6 +92,7 @@
 @import 'components/intro-story';
 @import 'components/loading-screen';
 @import 'components/modal-content';
+@import 'components/more-tools';
 @import 'components/navigator';
 @import 'components/network';
 @import 'components/voter-guide';
